requirements.txt
httpx
pydantic
loguru
pymongo  
python-dotenv
logging
pydantic_settings
click
socket
python-nmap
rich